Replace old serializer code with new one
